test: serve a real response to the radio stream connector
The connector waited out both of its connect timeouts because a listener
in the test's own process can only be accepted once connect() returns.
